Grite/requirements.txt
Django==3.1
django-bootstrap4==2.3.1
django-ckeditor==6.0.0
django-embed-video==1.4.0
django-js-asset==1.2.2
psycopg2==2.8.6


